APR-1855 - Move over Moderator repository functions into ModeratorRepository
