+++ src/SFA.DAS.ApplyService.Application.UnitTests/Handlers/GetModeratorPageReviewOutcomesForSectionHandlerTests/GetModeratorPageReviewOutcomesForSectionHandlerTests.cs
@@ -14,13 +14,13 @@ namespace SFA.DAS.ApplyService.Application.UnitTests.Handlers.GetModeratorPageRe
     [TestFixture]
     public class GetModeratorPageReviewOutcomesForSectionHandlerTests
     {
-        protected Mock<IAssessorRepository> _repository;
+        protected Mock<IModeratorRepository> _repository;
         protected GetModeratorPageReviewOutcomesForSectionHandler _handler;
 
         [SetUp]
         public void TestSetup()
         {
-            _repository = new Mock<IAssessorRepository>();
+            _repository = new Mock<IModeratorRepository>();
             _handler = new GetModeratorPageReviewOutcomesForSectionHandler(_repository.Object, Mock.Of<ILogger<GetModeratorPageReviewOutcomesForSectionHandler>>());
         }
