+++ src/SFA.DAS.ApplyService.Application/Apply/IAssessorRepository.cs
@@ -1,9 +1,7 @@
-
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using SFA.DAS.ApplyService.Domain.Apply.Assessor;
-using SFA.DAS.ApplyService.Domain.Apply.Moderator;
 
 namespace SFA.DAS.ApplyService.Application.Apply
 {
@@ -22,12 +20,6 @@ namespace SFA.DAS.ApplyService.Application.Apply
         Task<List<AssessorPageReviewOutcome>> GetAssessorPageReviewOutcomesForSection(Guid applicationId, int sequenceNumber, int sectionNumber, string userId);
         Task<List<AssessorPageReviewOutcome>> GetAllAssessorPageReviewOutcomes(Guid applicationId, string userId);
         Task UpdateAssessorReviewStatus(Guid applicationId, string userId, string status);
-
-        Task SubmitModeratorPageOutcome(Guid applicationId, int sequenceNumber, int sectionNumber, string pageId, string userId, string status, string comment, string externalComment);
-        Task<BlindAssessmentOutcome> GetBlindAssessmentOutcome(Guid applicationId, int sequenceNumber, int sectionNumber, string pageId);
-        Task<ModeratorPageReviewOutcome> GetModeratorPageReviewOutcome(Guid applicationId, int sequenceNumber, int sectionNumber, string pageId);
-        Task<List<ModeratorPageReviewOutcome>> GetModeratorPageReviewOutcomesForSection(Guid applicationId, int sequenceNumber, int sectionNumber);
-        Task<List<ModeratorPageReviewOutcome>> GetAllModeratorPageReviewOutcomes(Guid applicationId);
         Task CreateAssessorPageOutcomes(List<AssessorPageReviewOutcome> assessorPageReviewOutcomes);
     }
 }
