+++ src/SFA.DAS.ApplyService.Application/Apply/Moderator/GetAllModeratorPageReviewOutcomesHandler.cs
@@ -9,10 +9,10 @@ namespace SFA.DAS.ApplyService.Application.Apply.Moderator
 {
     public class GetAllModeratorPageReviewOutcomesHandler : IRequestHandler<GetAllModeratorPageReviewOutcomesRequest, List<ModeratorPageReviewOutcome>>
     {
-        private readonly IAssessorRepository _repository;
+        private readonly IModeratorRepository _repository;
         private readonly ILogger<GetAllModeratorPageReviewOutcomesHandler> _logger;
 
-        public GetAllModeratorPageReviewOutcomesHandler(IAssessorRepository repository, ILogger<GetAllModeratorPageReviewOutcomesHandler> logger)
+        public GetAllModeratorPageReviewOutcomesHandler(IModeratorRepository repository, ILogger<GetAllModeratorPageReviewOutcomesHandler> logger)
         {
             _repository = repository;
             _logger = logger;
