+++ src/SFA.DAS.ApplyService.Data/ModeratorRepository.cs
@@ -2,10 +2,13 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
+using System.Linq;
 using System.Threading.Tasks;
+using Dapper;
 using Microsoft.Extensions.Logging;
 using SFA.DAS.ApplyService.Application.Apply;
 using SFA.DAS.ApplyService.Configuration;
+using SFA.DAS.ApplyService.Domain.Apply;
 using SFA.DAS.ApplyService.Domain.Apply.Moderator;
 
 namespace SFA.DAS.ApplyService.Data
@@ -18,7 +21,166 @@ namespace SFA.DAS.ApplyService.Data
         public ModeratorRepository(IConfigurationService configurationService, ILogger<ModeratorRepository> logger)
         {
             _logger = logger;
-            _config = configurationService.GetConfig().Result;
+            _config = configurationService.GetConfig().GetAwaiter().GetResult();
+        }
+
+        public async Task<BlindAssessmentOutcome> GetBlindAssessmentOutcome(Guid applicationId, int sequenceNumber, int sectionNumber, string pageId)
+        {
+            using (var connection = new SqlConnection(_config.SqlConnectionString))
+            {
+                var blindAssessmentOutcomeResults = await connection.QueryAsync<BlindAssessmentOutcome>(
+                                                                @"SELECT outcome.[ApplicationId]
+			                                                            ,outcome.[SequenceNumber]
+			                                                            ,outcome.[SectionNumber]
+			                                                            ,outcome.[PageId]
+                                                                        ,apply.[Assessor1Name]
+			                                                            ,outcome.[Assessor1UserId]
+			                                                            ,outcome.[Assessor1ReviewStatus]
+			                                                            ,outcome.[Assessor1ReviewComment]
+                                                                        ,apply.[Assessor2Name]
+                                                                        ,outcome.[Assessor2UserId]
+			                                                            ,outcome.[Assessor2ReviewStatus]
+			                                                            ,outcome.[Assessor2ReviewComment]
+		                                                            FROM [dbo].[AssessorPageReviewOutcome] outcome
+                                                                    INNER JOIN [dbo].[Apply] apply ON outcome.ApplicationId = apply.ApplicationId
+		                                                            WHERE outcome.[ApplicationId] = @applicationId AND
+				                                                          outcome.[SequenceNumber] = @sequenceNumber AND
+				                                                          outcome.[SectionNumber] = @sectionNumber AND
+				                                                          outcome.[PageId] = @pageId AND
+                                                                          outcome.[Assessor1UserId] = apply.[Assessor1UserId] AND
+                                                                          outcome.[Assessor2UserId] = apply.[Assessor2UserId]",
+                    new { applicationId, sequenceNumber, sectionNumber, pageId });
+
+                return blindAssessmentOutcomeResults.FirstOrDefault();
+            }
+        }
+
+        public async Task<ModeratorPageReviewOutcome> GetModeratorPageReviewOutcome(Guid applicationId, int sequenceNumber, int sectionNumber, string pageId)
+        {
+            using (var connection = new SqlConnection(_config.SqlConnectionString))
+            {
+                var pageReviewOutcomeResults = await connection.QueryAsync<ModeratorPageReviewOutcome>(
+                                                                @"SELECT [ApplicationId]
+			                                                            ,[SequenceNumber]
+			                                                            ,[SectionNumber]
+			                                                            ,[PageId]
+			                                                            ,[ModeratorUserId] AS UserId
+			                                                            ,[ModeratorReviewStatus] AS [Status]
+			                                                            ,[ModeratorReviewComment] AS Comment
+                                                                        ,[ExternalComment]
+		                                                            FROM [dbo].[ModeratorPageReviewOutcome]
+		                                                            WHERE [ApplicationId] = @applicationId AND
+				                                                        [SequenceNumber] = @sequenceNumber AND
+				                                                        [SectionNumber] = @sectionNumber AND
+				                                                        [PageId] = @pageId",
+                    new { applicationId, sequenceNumber, sectionNumber, pageId });
+
+                return pageReviewOutcomeResults.FirstOrDefault();
+            }
+        }
+
+        public async Task<List<ModeratorPageReviewOutcome>> GetModeratorPageReviewOutcomesForSection(Guid applicationId, int sequenceNumber, int sectionNumber)
+        {
+            using (var connection = new SqlConnection(_config.SqlConnectionString))
+            {
+                var pageReviewOutcomeResults = await connection.QueryAsync<ModeratorPageReviewOutcome>(
+                                                                @"SELECT [ApplicationId]
+			                                                            ,[SequenceNumber]
+			                                                            ,[SectionNumber]
+			                                                            ,[PageId]
+			                                                            ,[ModeratorUserId] AS UserId
+			                                                            ,[ModeratorReviewStatus] AS [Status]
+			                                                            ,[ModeratorReviewComment] AS Comment
+                                                                        ,[ExternalComment]
+		                                                            FROM [dbo].[ModeratorPageReviewOutcome]
+		                                                            WHERE [ApplicationId] = @applicationId AND
+				                                                        [SequenceNumber] = @sequenceNumber AND
+				                                                        [SectionNumber] = @sectionNumber",
+                    new { applicationId, sequenceNumber, sectionNumber });
+
+                return pageReviewOutcomeResults.ToList();
+            }
+        }
+
+        public async Task<List<ModeratorPageReviewOutcome>> GetAllModeratorPageReviewOutcomes(Guid applicationId)
+        {
+            using (var connection = new SqlConnection(_config.SqlConnectionString))
+            {
+                var pageReviewOutcomeResults = await connection.QueryAsync<ModeratorPageReviewOutcome>(
+                                                                @"SELECT [ApplicationId]
+			                                                            ,[SequenceNumber]
+			                                                            ,[SectionNumber]
+			                                                            ,[PageId]
+			                                                            ,[ModeratorUserId] AS UserId
+			                                                            ,[ModeratorReviewStatus] AS [Status]
+			                                                            ,[ModeratorReviewComment] AS Comment
+                                                                        ,[ExternalComment]
+		                                                            FROM [dbo].[ModeratorPageReviewOutcome]
+		                                                            WHERE [ApplicationId] = @applicationId",
+                    new { applicationId });
+
+                return pageReviewOutcomeResults.ToList();
+            }
+        }
+
+        public async Task SubmitModeratorPageOutcome(Guid applicationId, int sequenceNumber, int sectionNumber, string pageId, string userId, string status, string comment, string externalComment)
+        {
+            using (var connection = new SqlConnection(_config.SqlConnectionString))
+            {
+                await connection.ExecuteAsync(
+                    @"IF NOT EXISTS (SELECT * FROM [ModeratorPageReviewOutcome]
+							         WHERE [ApplicationId] = @applicationId AND
+									        [SequenceNumber] = @sequenceNumber AND
+									        [SectionNumber] = @sectionNumber AND
+									        [PageId] = @pageId)
+		                BEGIN
+			                INSERT INTO [dbo].[ModeratorPageReviewOutcome]
+					                    ([ApplicationId]
+					                    ,[SequenceNumber]
+					                    ,[SectionNumber]
+					                    ,[PageId]
+					                    ,[ModeratorUserId]
+					                    ,[ModeratorReviewStatus]
+					                    ,[ModeratorReviewComment]
+                                        ,[ExternalComment]
+					                    ,[CreatedBy])
+				                    VALUES
+					                    (@applicationId
+					                    ,@sequenceNumber
+					                    ,@sectionNumber
+					                    ,@pageId
+					                    ,@userId
+					                    ,@status
+					                    ,@comment
+                                        ,@externalComment
+					                    ,@userId)                     
+		                END
+                      ELSE
+		                BEGIN
+			                UPDATE [ModeratorPageReviewOutcome]
+			                    SET [ModeratorUserId] = @userId
+                                    ,[ModeratorReviewStatus] = @status
+				                    ,[ModeratorReviewComment] = @comment
+                                    ,[ExternalComment] = @externalComment
+				                    ,[UpdatedAt] = GETUTCDATE()
+				                    ,[UpdatedBy] = @userId
+			                WHERE [ApplicationId] = @applicationId AND
+					                [SequenceNumber] = @sequenceNumber AND
+					                [SectionNumber] = @sectionNumber AND
+					                [PageId] = @pageId
+		                END",
+                    new { applicationId, sequenceNumber, sectionNumber, pageId, userId, status, comment, externalComment });
+
+                // APR-1633 - Update Moderation Status from 'New' to 'In Moderation'
+                await connection.ExecuteAsync(
+                    @"UPDATE [Apply]
+			            SET ModerationStatus = @inModerationStatus
+                            , UpdatedAt = GETUTCDATE()
+				            , UpdatedBy = @userId
+			            WHERE ApplicationId = @applicationId AND DeletedAt IS NULL
+                              AND ModerationStatus = @newStatus",
+                    new { applicationId, userId, inModerationStatus = ModerationStatus.InModeration, newStatus = ModerationStatus.New });
+            }
         }
 
         public async Task CreateModeratorPageOutcomes(List<ModeratorPageReviewOutcome> assessorPageReviewOutcomes)
